requirements.txt
torch==1.2.0
torchvision==0.4.0
numpy
PyYaml
scikit-learn
matplotlib
ray[dev]
jupyter
pandas
requests
psutil
tqdm
Pillow==6.2.0
gensim>=3.8.1
